requirements.txt
Flask==3.0.0
gunicorn==21.2.0
python-dotenv==1.2.1
psycopg2-binary>=2.9.0
openai>=1.0.0
requests>=2.28.0
snowflake-connector-python==3.5.0
pandas==2.3.3
numpy==2.2.6
SQLAlchemy==2.0.45
Jinja2==3.1.6
MarkupSafe==3.0.3
python-dateutil==2.9.0.post0
scikit-learn>=1.2.0
joblib>=1.2.0
openpyxl>=3.1.0
